use long values instead of int values

+++ projection/src/main/scala/org/apache/pekko/projection/r2dbc/internal/R2dbcOffsetStore.scala
@@ -486,7 +486,7 @@ private[projection] class R2dbcOffsetStore(
     }
   }
 
-  private def insertTimestampOffsetInTx(conn: Connection, records: immutable.IndexedSeq[Record]): Future[Int] = {
+  private def insertTimestampOffsetInTx(conn: Connection, records: immutable.IndexedSeq[Record]): Future[Long] = {
     def bindRecord(stmt: Statement, record: Record): Statement = {
       val slice = persistenceExt.sliceForPersistenceId(record.pid)
       val minSlice = timestampOffsetBySlicesSourceProvider.minSlice
@@ -796,7 +796,7 @@ private[projection] class R2dbcOffsetStore(
     }
   }
 
-  def deleteOldTimestampOffsets(): Future[Int] = {
+  def deleteOldTimestampOffsets(): Future[Long] = {
     if (idle.getAndSet(true)) {
       // no new offsets stored since previous delete
       Future.successful(0)
@@ -833,7 +833,7 @@ private[projection] class R2dbcOffsetStore(
           result.foreach { rows =>
             logger.debug(
               "Deleted [{}] timestamp offset rows until [{}] for projection [{}].",
-              rows: java.lang.Integer,
+              rows: java.lang.Long,
               until,
               projectionId.id)
           }
@@ -873,7 +873,7 @@ private[projection] class R2dbcOffsetStore(
     }
   }
 
-  private def deleteNewTimestampOffsetsInTx(conn: Connection, timestamp: Instant): Future[Int] = {
+  private def deleteNewTimestampOffsetsInTx(conn: Connection, timestamp: Instant): Future[Long] = {
     val currentState = getState()
     if (timestamp.isAfter(currentState.latestTimestamp)) {
       // nothing to delete
@@ -896,7 +896,7 @@ private[projection] class R2dbcOffsetStore(
         result.foreach { rows =>
           logger.debug(
             "Deleted [{}] timestamp offset rows >= [{}] for projection [{}].",
-            rows: java.lang.Integer,
+            rows: java.lang.Long,
             timestamp,
             projectionId.id)
         }
